+++ R/matrix_plot.R
@@ -1,6 +1,6 @@
 to_char_uneval_matrix <- function(x) {
   ex <- unlist(lapply(x, function(y) deparse(y$expr)))
-  ex[ex == "C"] <- ""
+  ex[ex == "0"] <- ""
   matrix(ex,
          byrow = TRUE,
          ncol = get_matrix_order(x),